Create synchronous handler for /stats command
Note, the method still makes use of async functions, but the request
to Discord to responded to synchronously. In this future, this will
be updated to use a DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE and a
callback to Discord once the data has processed.

Currently, a throttling exception from Discord will result in a
failure to process the message.